+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.0
 
+- Added Validation for the song length to be <45s
 - Added Validation for "All lanes must be used" rule
 - Added Validation for overlapping and duplicated notes
 - Added Validation for "mapped activity" percentage >75%